Change usertasks button to html <button> tag
Since the link is not actually a link, it is not linking to another page
nor to an anchor on the current page, change it to a button. This
represents the actual function better and has the positive side effect
that the button is now reachable via tab-key on the keyboard.

SPR-977